refactor: magic number 제거 및 에러 메시지 도메인별 상수화
